+++ backend/SDIA/Application/DTOs/UserDto.cs
@@ -1,4 +1,4 @@
-using Domain.Constants;
+using Domain.Constants.Enums;
 
 namespace Application.DTOs;
