+++ backend/SDIA/Application/Entities/Response/DocumentClassificationResponse.cs
@@ -1,4 +1,4 @@
-using Domain.Constants;
+using Domain.Constants.Enums;
 
 namespace Application.Entities.Response;
