+++ backend/SDIA/Application/Handlers/User/GetUserProfileByIdHandler.cs
@@ -1,7 +1,7 @@
 using Application.DTOs;
 using Application.Interfaces;
 using Application.Query.User;
-using Domain.Constants;
+using Domain.Constants.Enums;
 using Domain.Exceptions;
 using Mapster;
 using MediatR;
