+++ backend/SDIA/Application/Handlers/User/LoginUserHandler.cs
@@ -1,6 +1,6 @@
 using Application.Commands.User;
 using Application.Interfaces;
-using Domain.Constants;
+using Domain.Constants.Enums;
 using Domain.Exceptions;
 using MediatR;
 using Microsoft.EntityFrameworkCore;
