+++ backend/SDIA/Domain/Constants/Enums/DocumentType.cs
@@ -1,6 +1,6 @@
 using System.Text.Json.Serialization;
 
-namespace Domain.Constants;
+namespace Domain.Constants.Enums;
 
 public enum DocumentType
 {
