+++ backend/SDIA/Domain/Constants/Enums/FolderType.cs
@@ -1,0 +1,8 @@
+namespace Domain.Constants.Enums;
+
+public enum FolderType
+{
+    CarFromAnotherCountry,
+    CarNeverRegistered,
+    AlreadyRegisteredVehicleInCountry,
+}
