--- backend/SDIA/Domain/Constants/enums/FolderType.cs
@@ -1,6 +1,0 @@
-namespace Domain.Constants;
-
-public enum FolderType
-{
-    AlreadyRegisteredVehicleInCountry
-}
