+++ backend/SDIA/Domain/Entities/Folder.cs
@@ -1,6 +1,6 @@
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
-using Domain.Constants;
+using Domain.Constants.Enums;
 using Domain.Interfaces;
 
 namespace Domain.Entities;
