+++ backend/SDIA/Presentation/Entities/Document/Responses/DocumentInfoResponse.cs
@@ -1,4 +1,4 @@
-using Domain.Constants;
+using Domain.Constants.Enums;
 
 namespace SDIA.Entities.Document.Responses;
